할인 금액 계산에 대한 모든 세부사항은 할인 정책에서 결정한다
