+++ src/main/java/com/namu/object/_05_movie/AmountDiscountPolicy.java
@@ -1,16 +1,18 @@
 package com.namu.object._05_movie;
 
+import com.namu.object._02_movie.DefaultDiscountPolicy;
 import com.namu.object._02_movie.Money;
 
 public class AmountDiscountPolicy extends DiscountPolicy {
     private Money discountAmount;
 
-    public AmountDiscountPolicy(Money discountAmount) {
+    public AmountDiscountPolicy(Money discountAmount, DiscountCondition... discountConditions) {
+        super(discountConditions);
         this.discountAmount = discountAmount;
     }
 
     @Override
-    Money getDiscountAmount(Movie movie) {
+    protected Money getDiscountAmount(Movie movie) {
         return discountAmount;
     }
