+++ src/main/java/com/namu/object/_05_movie/NoneDiscountPolicy.java
@@ -1,11 +1,11 @@
 package com.namu.object._05_movie;
 
+import com.namu.object._02_movie.DefaultDiscountPolicy;
 import com.namu.object._02_movie.Money;
 
 public class NoneDiscountPolicy extends DiscountPolicy {
-
     @Override
-    Money getDiscountAmount(Movie movie) {
+    protected Money getDiscountAmount(Movie movie) {
         return Money.ZERO;
     }
 }
